# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "LinkedIn",
            "url": "https://www.linkedin.com/in/patricio-ernesto-antonio-arena-08a0a9133/",
            "email": "patricio.e.arena@gmail.com"
        }
    },
    "paths": {
        "/game/all": {
            "get": {
                "description": "Get a list of all games",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "List games",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Game"
                            }
                        }
                    }
                }
            }
        },
        "/game/all/order-by-discount": {
            "get": {
                "description": "Get a list of all games order by discount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "List games order by discount",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Game"
                            }
                        }
                    }
                }
            }
        },
        "/game/page/": {
            "get": {
                "description": "Get a list of all games with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "List games with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "number",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit results",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Game"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Game": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "discount": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price_with_discounted": {
                    "type": "string"
                },
                "price_without_discounted": {
                    "type": "string"
                }
            }
        }
    }
}